requirements.txt
Flask==2.2.3
Werkzeug==2.2.3
requests==2.28.2
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
python-dotenv==1.0.0
Flask-Migrate==4.0.7 
pytz
notion-client